backend/requirements.txt
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
cryptography==44.0.3
daphne==4.1.2
decorator==5.2.1
distlib==0.3.9
Django==5.0.14
django-cors-headers==4.7.0
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
et_xmlfile==2.0.0
executing==2.2.0
filelock==3.18.0
Flask==3.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-ui-poll==1.0.0
jupyterlab_widgets==3.0.15
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
MouseInfo==0.1.3
msal==1.32.3
msgpack==1.1.1
mssql-django==1.5
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
parso==0.8.4
pillow==11.2.1
pipenv==2025.0.1
platformdirs==4.3.7
powerbiclient==3.1.1
prompt_toolkit==3.0.51
pure_eval==0.2.3
py4j==0.10.9.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
Pygments==2.19.1
PyJWT==2.9.0
PyMsgBox==1.0.9
pyodbc==5.2.0
pyOpenSSL==25.0.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pyspark==3.5.5
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-telegram-bot==22.2
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
redis==6.2.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
service-identity==24.2.0
setuptools==80.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.35.0
virtualenv==20.30.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
whitenoise==6.9.0
widgetsnbextension==4.0.14
xlrd==2.0.1
XlsxWriter==3.2.3
zope.interface==7.2
